+++ examples/website/wms/app.tsx
@@ -7,10 +7,39 @@ import {StaticMap} from 'react-map-gl';
 import DeckGL from '@deck.gl/react/typed';
 import {MapController} from '@deck.gl/core/typed';
 import {ImageryLayer} from './imagery-layer';
-import type {_ImageSourceMetadata as ImageSourceMetadata} from '@loaders.gl/wms';
+import type {ImageSourceMetadata} from '@loaders.gl/wms';
 
 import ControlPanel from './components/control-panel';
 import {INITIAL_CATEGORY_NAME, INITIAL_EXAMPLE_NAME, INITIAL_MAP_STYLE, EXAMPLES} from './examples';
+
+import {_getArcGISServices, CSWService} from '@loaders.gl/wms';
+
+const catalogService = new CSWService({url: 'https://gamone.whoi.edu/csw'}); // https://deims.org/pycsw/catalogue'});
+const services = await catalogService.getServiceDirectory();
+console.log(JSON.stringify(services, null, 2));
+
+const examples = EXAMPLES['gamone'] = {};
+let i = 0;
+for (const service of services) {
+  examples[`${service.name.replace('-', '')}`] = {
+    service: service.url,
+    serviceType: 'wms',
+    layers: ['THREDDS'],
+    viewState: {
+      longitude: -122.4,
+      latitude: 37.74,
+      zoom: 9,
+      minZoom: 1,
+      maxZoom: 20,
+      pitch: 0,
+      bearing: 0    
+    }
+  }
+}
+
+
+// const services = await _getArcGISServices('https://sampleserver6.arcgisonline.com/arcgis/rest/services'); // /Water_Network_Base_Map/MapServer
+// console.log(JSON.stringify(services, null, 2));
 
 export const INITIAL_VIEW_STATE = {
   latitude: 49.254,